requirements.txt
alembic==1.13.3
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-restx==1.3.0
greenlet==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow2==2.20.2
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
